get_hidden_states function added
get_hidden_states function added in the utils.utils file. it takes a model and its tokenizer and returns the output text, output tokens and hidden states of every layer. Also added a new model in generator.clients file, llama3_2_hd (which is llama3.2 and uses this function to return text and hidden states).